+++ src/ReplicatedStorage/Animations/Package/Util/AnimationIdsUtil.lua
@@ -5,8 +5,9 @@ function AnimationIdsUtil.HasProperties(
 	propertiesSettings: {
         Priority: Enum.AnimationPriority?,
         Looped: boolean?,
-        MarkerTimes: boolean?, -- Set this to true for animations that should have accessible animation marker times
-		DoUnpack: boolean?, -- Sets the key value pairs for everything up one level instead of in the passed animation ids able
+        StartSpeed: number?,
+        MarkerTimes: boolean?,
+		DoUnpack: boolean?
 	}
 ): {}
     local doUnpack = propertiesSettings.DoUnpack
@@ -31,7 +32,7 @@ function AnimationIdsUtil.HasAnimatedObject(
 	animatedObjectSettings: {
 		AutoAttach: boolean?,
 		AutoDetach: boolean?,
-		DoUnpack: boolean?, -- Sets the key value pairs for everything up one level instead of in the passed animation ids able
+		DoUnpack: boolean?
 	}
 ): {}
     local doUnpack = animatedObjectSettings.DoUnpack